+++ src/plugins/aws/ssm/index.ts
@@ -9,11 +9,11 @@ This file is part of @p0security/cli
 You should have received a copy of the GNU General Public License along with @p0security/cli. If not, see <https://www.gnu.org/licenses/>.
 **/
 import {
-  PRIVATE_KEY_PATH,
   ScpCommandArgs,
   SshCommandArgs,
   SshRequest,
 } from "../../../commands/shared";
+import { PRIVATE_KEY_PATH } from "../../../common/keys";
 import { print2 } from "../../../drivers/stdio";
 import { Authn } from "../../../types/identity";
 import { assumeRoleWithOktaSaml } from "../../okta/aws";
