+++ src/plugins/ssh-agent/index.ts
@@ -8,7 +8,7 @@ This file is part of @p0security/cli
 
 You should have received a copy of the GNU General Public License along with @p0security/cli. If not, see <https://www.gnu.org/licenses/>.
 **/
-import { PRIVATE_KEY_PATH } from "../../commands/shared";
+import { PRIVATE_KEY_PATH } from "../../common/keys";
 import { print2 } from "../../drivers/stdio";
 import { AgentArgs } from "./types";
 import { SpawnOptionsWithoutStdio, spawn } from "node:child_process";
